# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Budget": {
        "properties": {
          "daily_spend_cap_cents": {
            "title": "Daily Spend Cap Cents",
            "type": "integer"
          },
          "remaining_today_cents": {
            "title": "Remaining Today Cents",
            "type": "integer"
          },
          "spent_today_cents": {
            "title": "Spent Today Cents",
            "type": "integer"
          }
        },
        "required": [
          "daily_spend_cap_cents",
          "spent_today_cents",
          "remaining_today_cents"
        ],
        "title": "Budget",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OptionsSummary": {
        "description": "Static configuration + live budget for driving frontend forms.",
        "properties": {
          "aspect_ratios": {
            "items": {
              "type": "string"
            },
            "title": "Aspect Ratios",
            "type": "array"
          },
          "daily_spend_cap_cents": {
            "title": "Daily Spend Cap Cents",
            "type": "integer"
          },
          "default_image_size": {
            "title": "Default Image Size",
            "type": "string"
          },
          "default_model": {
            "title": "Default Model",
            "type": "string"
          },
          "image_sizes": {
            "items": {
              "type": "string"
            },
            "title": "Image Sizes",
            "type": "array"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "title": "Models",
            "type": "array"
          },
          "ref_image_roles": {
            "items": {
              "type": "string"
            },
            "title": "Ref Image Roles",
            "type": "array"
          },
          "ref_image_weight_explanation": {
            "title": "Ref Image Weight Explanation",
            "type": "string"
          },
          "ref_set_immutability_explanation": {
            "title": "Ref Set Immutability Explanation",
            "type": "string"
          },
          "remaining_today_cents": {
            "title": "Remaining Today Cents",
            "type": "integer"
          },
          "scene_immutability_explanation": {
            "title": "Scene Immutability Explanation",
            "type": "string"
          },
          "spent_today_cents": {
            "title": "Spent Today Cents",
            "type": "integer"
          }
        },
        "required": [
          "models",
          "image_sizes",
          "aspect_ratios",
          "ref_image_roles",
          "default_model",
          "default_image_size",
          "daily_spend_cap_cents",
          "spent_today_cents",
          "remaining_today_cents",
          "ref_image_weight_explanation",
          "ref_set_immutability_explanation",
          "scene_immutability_explanation"
        ],
        "title": "OptionsSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "LoreCraft3000",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/budget": {
      "get": {
        "operationId": "budget_api_v1_budget_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Timezone",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Timezone"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Budget"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Budget",
        "tags": [
          "api-v1",
          "api-v1-options"
        ]
      }
    },
...
    "/api/v1/options/summary": {
      "get": {
        "operationId": "options_summary_api_v1_options_summary_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Timezone",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Timezone"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Options Summary",
        "tags": [
          "api-v1",
          "api-v1-options"
        ]
      }
    },
...
  }
}